% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_calculate_pnl.R
\name{.calculate_unrealized_pnl}
\alias{.calculate_unrealized_pnl}
\title{Calculate Unrealized PnL}
\usage{
.calculate_unrealized_pnl(
  price,
  long_avg_entry_price,
  short_avg_entry_price,
  long_notional,
  short_notional,
  fee_rate
)
}
\description{
Internal function to update unrealized PnL
}
\keyword{internal}
