% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_agents.R
\name{sim_agent_rankings}
\alias{sim_agent_rankings}
\title{Compute current agent rankings}
\usage{
sim_agent_rankings(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A data.table of agent rankings.
}
\description{
Rankings combine current per-agent account equity with order activity.
}
