% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_assets.R
\name{sim_assets}
\alias{sim_assets}
\title{List registered exchange assets}
\usage{
sim_assets(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A data.table of assets.
}
\description{
List registered exchange assets
}
