% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bond_schedules.R
\name{sim_bond_schedules}
\alias{sim_bond_schedules}
\title{List durable bond schedules}
\usage{
sim_bond_schedules(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A data.table of bond schedule state.
}
\description{
List durable bond schedules
}
