% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exchange.R
\name{sim_exchange_account}
\alias{sim_exchange_account}
\title{Get simulated exchange account state}
\usage{
sim_exchange_account(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A one-row data.table with the latest account snapshot.
}
\description{
Get simulated exchange account state
}
