% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exchange.R
\name{sim_exchange_add_bars}
\alias{sim_exchange_add_bars}
\title{Append market bars to a simulated exchange}
\usage{
sim_exchange_add_bars(exchange, bars)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{bars}{Market bars coercible by \code{as_market_bars()}.}
}
\value{
The exchange, invisibly.
}
\description{
Append market bars to a simulated exchange
}
