% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exchange.R
\name{sim_exchange_export_events}
\alias{sim_exchange_export_events}
\title{Export exchange simulation events}
\usage{
sim_exchange_export_events(exchange, path, format = c("csv", "fst"))
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{path}{Output directory.}

\item{format}{File format, either \code{csv} or \code{fst}.}
}
\value{
Invisibly returns written file paths.
}
\description{
Export exchange simulation events
}
