% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exchange.R
\name{sim_exchange_new_events}
\alias{sim_exchange_new_events}
\title{Get new events since the previous exchange run}
\usage{
sim_exchange_new_events(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A data.table of newly observed simulation events.
}
\description{
Get new events since the previous exchange run
}
