% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exchange.R
\name{sim_exchange_orders}
\alias{sim_exchange_orders}
\title{Get simulated exchange orders}
\usage{
sim_exchange_orders(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A data.table of accepted/cancelled intent-level orders.
}
\description{
Get simulated exchange orders
}
