% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exchange.R
\name{sim_exchange_positions}
\alias{sim_exchange_positions}
\title{Get simulated exchange positions}
\usage{
sim_exchange_positions(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A one-row data.table with the latest position snapshot.
}
\description{
Get simulated exchange positions
}
