% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_feed.R
\name{sim_feed_start}
\alias{sim_feed_start}
\title{Start a configured live feed}
\usage{
sim_feed_start(exchange, now = Sys.time())
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{now}{Current time used to initialize the schedule.}
}
\value{
Feed status.
}
\description{
Start a configured live feed
}
