% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_feed.R
\name{sim_feed_status}
\alias{sim_feed_status}
\title{Get live feed status}
\usage{
sim_feed_status(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A list describing feed state.
}
\description{
Get live feed status
}
