% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_feed.R
\name{sim_feed_stop}
\alias{sim_feed_stop}
\title{Stop a configured live feed}
\usage{
sim_feed_stop(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
Feed status.
}
\description{
Stop a configured live feed
}
