% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ledgers.R
\name{sim_fills}
\alias{sim_fills}
\title{Extract fill events from a simulation}
\usage{
sim_fills(sim)
}
\arguments{
\item{sim}{A simulation result returned by \code{sim_backtest()}.}
}
\value{
A data.table of filled trade events.
}
\description{
Extract fill events from a simulation
}
