% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_service.R
\name{sim_live_service_run}
\alias{sim_live_service_run}
\title{Run a local live exchange service}
\usage{
sim_live_service_run(
  exchange = sim_exchange_new(),
  host = "127.0.0.1",
  port = 8080
)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{host}{Host interface.}

\item{port}{Port.}
}
\value{
The result of \code{plumber}'s \code{run()} method.
}
\description{
Run a local live exchange service
}
