% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dashboard.R
\name{sim_live_state_dashboard_open}
\alias{sim_live_state_dashboard_open}
\title{Open a live-state dashboard}
\usage{
sim_live_state_dashboard_open(
  exchange = sim_exchange_new(),
  path = tempfile("tradesimr-live-state-")
)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{path}{Output directory.}
}
\value{
Invisibly returns the dashboard index path.
}
\description{
Open a live-state dashboard
}
