% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_feed.R
\name{sim_market_model_configure}
\alias{sim_market_model_configure}
\title{Configure a market-level simulation model}
\usage{
sim_market_model_configure(exchange, config = sim_market_model_config())
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{config}{A list from \code{sim_market_model_config()}.}
}
\value{
The market model configuration, invisibly.
}
\description{
Configure a market-level simulation model
}
