% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ledgers.R
\name{sim_read_account}
\alias{sim_read_account}
\title{Read exported account snapshots}
\usage{
sim_read_account(path)
}
\arguments{
\item{path}{Export directory created by \code{sim_export()}.}
}
\value{
A data.table of account snapshots.
}
\description{
Read exported account snapshots
}
