% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ledgers.R
\name{sim_read_events}
\alias{sim_read_events}
\title{Read exported simulation events}
\usage{
sim_read_events(path)
}
\arguments{
\item{path}{Export directory created by \code{sim_export()}.}
}
\value{
A data.table of events.
}
\description{
Read exported simulation events
}
