% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_schema.R
\name{sim_schema_version}
\alias{sim_schema_version}
\title{Get the tradesimr schema version}
\usage{
sim_schema_version()
}
\value{
A scalar character schema version.
}
\description{
Get the tradesimr schema version
}
