% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_schema.R
\name{sim_schemas}
\alias{sim_schemas}
\title{Simulation table schemas}
\usage{
sim_schemas()
}
\value{
A named list of empty data.tables representing durable simulation
table schemas.
}
\description{
Simulation table schemas
}
