% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_agents.R
\name{sim_strategy_list}
\alias{sim_strategy_list}
\title{List registered strategy ids}
\usage{
sim_strategy_list(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
Character vector of registered strategy ids.
}
\description{
List registered strategy ids
}
