% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_calendars.R
\name{sim_trading_calendars}
\alias{sim_trading_calendars}
\title{List built-in trading calendars}
\usage{
sim_trading_calendars()
}
\value{
A data.table describing the deterministic built-in session rules.
}
\description{
List built-in trading calendars
}
