% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_schema.R
\name{TRADESIMR_ACCOUNT_SCHEMA_VERSION}
\alias{TRADESIMR_ACCOUNT_SCHEMA_VERSION}
\title{Heterogeneous account schema version}
\usage{
TRADESIMR_ACCOUNT_SCHEMA_VERSION
}
\description{
Heterogeneous account schema version
}
