% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_schema.R
\docType{data}
\name{TRADESIMR_SCHEMA_VERSION}
\alias{TRADESIMR_SCHEMA_VERSION}
\title{tradesimr durable schema version}
\format{
An object of class \code{character} of length 1.
}
\usage{
TRADESIMR_SCHEMA_VERSION
}
\description{
tradesimr durable schema version
}
\keyword{datasets}
