% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_sim.R
\name{vec_sim_gen_plot}
\alias{vec_sim_gen_plot}
\title{Plot vectorized simulation results}
\usage{
vec_sim_gen_plot(vec_sim_res, report_mode = c("simple", "full"))
}
\arguments{
\item{vec_sim_res}{Result from \code{vec_sim_run_backtest()}.}

\item{report_mode}{Plot report mode.}
}
\description{
Plot vectorized simulation results
}
