% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_sim.R
\name{vec_sim_gen_summary}
\alias{vec_sim_gen_summary}
\title{Summarize vectorized simulation results}
\usage{
vec_sim_gen_summary(
  vec_sim_res,
  report_mode = c("short_txt", "long_txt", "dt")
)
}
\arguments{
\item{vec_sim_res}{Result from \code{vec_sim_run_backtest()}.}

\item{report_mode}{Summary mode.}
}
\value{
Text or a one-row data.table depending on \code{report_mode}.
}
\description{
Summarize vectorized simulation results
}
