% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_sim.R
\name{vec_sim_gen_summary_table}
\alias{vec_sim_gen_summary_table}
\title{Summarize a batch of vectorized simulations}
\usage{
vec_sim_gen_summary_table(DTs)
}
\arguments{
\item{DTs}{List of vectorized simulation data.tables.}
}
\value{
A data.table of summaries.
}
\description{
Summarize a batch of vectorized simulations
}
