% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaperTrader.R
\name{PaperTrader}
\alias{PaperTrader}
\title{Paper trader demo client}
\description{
Legacy R6 demonstration client for \code{PaperTradingPlatform}. The production
simulation path is \code{sim_backtest()} and the C++ execution engine.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PaperTrader-initialize}{\code{PaperTrader$new()}}
    \item \href{#method-PaperTrader-place_order}{\code{PaperTrader$place_order()}}
    \item \href{#method-PaperTrader-cancel_order}{\code{PaperTrader$cancel_order()}}
    \item \href{#method-PaperTrader-get_all_orders}{\code{PaperTrader$get_all_orders()}}
    \item \href{#method-PaperTrader-get_order}{\code{PaperTrader$get_order()}}
    \item \href{#method-PaperTrader-get_wallet}{\code{PaperTrader$get_wallet()}}
    \item \href{#method-PaperTrader-get_position}{\code{PaperTrader$get_position()}}
    \item \href{#method-PaperTrader-clone}{\code{PaperTrader$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-initialize}{}}}
\subsection{\code{PaperTrader$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$new(user_id, platform)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-place_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-place_order}{}}}
\subsection{\code{PaperTrader$place_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$place_order(inst_id, type, pos, size, price, pricing_method, tag)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-cancel_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-cancel_order}{}}}
\subsection{\code{PaperTrader$cancel_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$cancel_order(order_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-get_all_orders"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-get_all_orders}{}}}
\subsection{\code{PaperTrader$get_all_orders()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$get_all_orders()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-get_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-get_order}{}}}
\subsection{\code{PaperTrader$get_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$get_order(order_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-get_wallet"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-get_wallet}{}}}
\subsection{\code{PaperTrader$get_wallet()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$get_wallet()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-get_position}{}}}
\subsection{\code{PaperTrader$get_position()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$get_position()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTrader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTrader-clone}{}}}
\subsection{\code{PaperTrader$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTrader$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
