% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaperTradingPlatform.R
\name{PaperTradingPlatform}
\alias{PaperTradingPlatform}
\title{Paper trading platform demo}
\description{
Legacy R6 demonstration of a paper trading platform. The production
simulation path is \code{sim_backtest()} and the C++ execution engine.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PaperTradingPlatform-register_user}{\code{PaperTradingPlatform$register_user()}}
    \item \href{#method-PaperTradingPlatform-place_user_order}{\code{PaperTradingPlatform$place_user_order()}}
    \item \href{#method-PaperTradingPlatform-cancel_user_order}{\code{PaperTradingPlatform$cancel_user_order()}}
    \item \href{#method-PaperTradingPlatform-get_user_orders}{\code{PaperTradingPlatform$get_user_orders()}}
    \item \href{#method-PaperTradingPlatform-get_user_order}{\code{PaperTradingPlatform$get_user_order()}}
    \item \href{#method-PaperTradingPlatform-get_live_orders}{\code{PaperTradingPlatform$get_live_orders()}}
    \item \href{#method-PaperTradingPlatform-get_order}{\code{PaperTradingPlatform$get_order()}}
    \item \href{#method-PaperTradingPlatform-update_bar}{\code{PaperTradingPlatform$update_bar()}}
    \item \href{#method-PaperTradingPlatform-process_order}{\code{PaperTradingPlatform$process_order()}}
    \item \href{#method-PaperTradingPlatform-fill_order}{\code{PaperTradingPlatform$fill_order()}}
    \item \href{#method-PaperTradingPlatform-get_user_wallet}{\code{PaperTradingPlatform$get_user_wallet()}}
    \item \href{#method-PaperTradingPlatform-get_user_position}{\code{PaperTradingPlatform$get_user_position()}}
    \item \href{#method-PaperTradingPlatform-update_user_wallet}{\code{PaperTradingPlatform$update_user_wallet()}}
    \item \href{#method-PaperTradingPlatform-update_user_position}{\code{PaperTradingPlatform$update_user_position()}}
    \item \href{#method-PaperTradingPlatform-clone}{\code{PaperTradingPlatform$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-register_user"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-register_user}{}}}
\subsection{\code{PaperTradingPlatform$register_user()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$register_user(user_id, initial_balance = 10000)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-place_user_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-place_user_order}{}}}
\subsection{\code{PaperTradingPlatform$place_user_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$place_user_order(
  user_id,
  inst_id,
  type,
  pos,
  size,
  price,
  pricing_method,
  tag
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-cancel_user_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-cancel_user_order}{}}}
\subsection{\code{PaperTradingPlatform$cancel_user_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$cancel_user_order(user_id, order_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-get_user_orders"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-get_user_orders}{}}}
\subsection{\code{PaperTradingPlatform$get_user_orders()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$get_user_orders(user_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-get_user_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-get_user_order}{}}}
\subsection{\code{PaperTradingPlatform$get_user_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$get_user_order(user_id, order_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-get_live_orders"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-get_live_orders}{}}}
\subsection{\code{PaperTradingPlatform$get_live_orders()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$get_live_orders()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-get_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-get_order}{}}}
\subsection{\code{PaperTradingPlatform$get_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$get_order(order_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-update_bar"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-update_bar}{}}}
\subsection{\code{PaperTradingPlatform$update_bar()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$update_bar(inst_id, timestamp, open, high, low, close)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-process_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-process_order}{}}}
\subsection{\code{PaperTradingPlatform$process_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$process_order(order)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-fill_order"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-fill_order}{}}}
\subsection{\code{PaperTradingPlatform$fill_order()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$fill_order(order_id, fill_price)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-get_user_wallet"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-get_user_wallet}{}}}
\subsection{\code{PaperTradingPlatform$get_user_wallet()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$get_user_wallet(user_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-get_user_position"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-get_user_position}{}}}
\subsection{\code{PaperTradingPlatform$get_user_position()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$get_user_position(user_id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-update_user_wallet"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-update_user_wallet}{}}}
\subsection{\code{PaperTradingPlatform$update_user_wallet()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$update_user_wallet(user_id, delta_wallet_balance)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-update_user_position"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-update_user_position}{}}}
\subsection{\code{PaperTradingPlatform$update_user_position()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$update_user_position(order)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaperTradingPlatform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaperTradingPlatform-clone}{}}}
\subsection{\code{PaperTradingPlatform$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PaperTradingPlatform$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
