% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exchange.R
\name{sim_exchange_step}
\alias{sim_exchange_step}
\title{Step a simulated exchange with one or more bars}
\usage{
sim_exchange_step(exchange, bars)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{bars}{Market bars coercible by \code{as_market_bars()}.}
}
\value{
The incremental simulation snapshots.
}
\description{
Step a simulated exchange with one or more bars
}
