% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_feed.R
\name{sim_feed_configure}
\alias{sim_feed_configure}
\title{Configure a live exchange feed}
\usage{
sim_feed_configure(exchange, config = sim_feed_config())
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}

\item{config}{Feed configuration list. A list with \code{configs} may be used to
configure all registered assets in one call.}
}
\value{
The feed configuration, invisibly.
}
\description{
Configure a live exchange feed
}
