% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_feed.R
\name{sim_market_model_status}
\alias{sim_market_model_status}
\title{Get market-level simulation model status}
\usage{
sim_market_model_status(exchange)
}
\arguments{
\item{exchange}{A \code{tradesimr_exchange}.}
}
\value{
A list describing the market-level simulation model.
}
\description{
Get market-level simulation model status
}
